Ajuste de carpetas controlador
Se elimina la carpeta CONTROLLER y se agrupan todos estos dentro de CONTROLADOR
